Remove duplicate child components

+++ packages/subgraph/start-local.sh
@@ -1,5 +1,14 @@
 set -e
 
+root_dir=$(dirname $(dirname $(dirname $(readlink -f "$0"))))
+env_path="${root_dir}/.env.local"
+source "${env_path}"
+
+if [ "$REACT_APP_SENIOR_POOL_PAGE_VERSION"!="V2" ]; then
+    echo "Subgraph feature flag is toggled off. Exiting..."
+    exit 0
+fi
+
 docker-compose down -v;
 
 echo "Building the sugraph..."
@@ -26,4 +35,4 @@ npm run build
 
 npm run create-local
 
-npm run deploy-local --version-label v0.0.1
+npm run deploy-local
